Core code cleanup (#206)
* add delay function
* todo about delay_isr
* remove arduino defines and fix all apps to use core-api/hal-api
* delay for local target
* remove warnings of task_equal
* fix BSP_SD_Init
* fix USBD_static
* grio read constant pointer to gpio
* add TODO about ISR context
* const void* arg for pubsub api
* mark unused functions
* app pointers now pointed to constant apps
* fix printf format
* fix "unused" warnings in local target
* fix const pin read in local target
* fix int to pointer warnings in local target
* power read mutex error fix
* delete old makefile
* add -werror

+++ applications/lf-rfid/lf-rfid.c
@@ -28,7 +28,7 @@ static void render_callback(CanvasApi* canvas, void* ctx) {
 
     canvas->draw_str(canvas, 2, 24, state->on ? "ON" : "OFF");
     char buf[12];
-    sprintf(buf, "%d kHz", state->freq_khz);
+    sprintf(buf, "%d kHz", (int)state->freq_khz);
     canvas->draw_str(canvas, 2, 36, buf);
 
     release_mutex((ValueMutex*)ctx, state);
